backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.25.2
python-dotenv==1.0.1
itsdangerous==2.2.0
pydantic==2.10.6
email-validator==2.2.0
pytest==8.2.2
PyJWT==2.8.0

# Supabase and PostgreSQL
asyncpg==0.29.0
supabase==2.7.4

# OpenAI
openai==1.55.3

# Vector database (optional - for RAG)
pgvector==0.3.6

# Additional utilities
python-multipart==0.0.7  # For file uploads
pydantic-settings==2.6.1

